wip: support 'ask' for clipboard permissions
WebEngine now supports prompting for permission when a web page tried to
access the clipboard. Previously we only supported fixed permissions that
applied globally. This commit is a bit rough currently, things we might want
to fine tune:

content.javascript.clipboard and BoolAsk

All the existing settings that support permissions prompts are of type
BoolAsk, the clipboard one is of type "string". Currently I've just added
"ask" to the allowed set of strings and added a sanity check to
`feature_permission()` so that it doesn't auto-approve for non-bool truthy
values. But this permission request signal does get called on 6.8 in some
cases so it's printing out the warning in some tests. That might be okay, or
maybe we could do something around changing the type of the clipboard setting
to have a to_bool() method or something.

tests on Qt < 6.8

The new prompt tests require QtWebEngine 6.8 (compiled) to work. They aren't
checking for that currently.

Given I have a fresh instance

This isn't required on 6.8 (not compiled) when we are setting AskEveryTime for
the new persistent permissions policy.