Change check for new API to try/catch
Previously it would have crashed with an AttributeError if a user had a
PyQt of 6.8 but Qt of 6.7.

Switch to catching any AttributeError which will hopefully cover
either or both of Qt and PyQt not being a new enough version, even if
it's a bit less of an explicit check.