tests: Stabilize reloading in tests
The test added in fbd148f98399090704d2cb055c928ca14f18cba2 was flaky because
reloading didn't wait for the page load to actually finish fully, and thus the
element not being found on the page.

Fix this by providing the page path to the step, and making sure it waits for it
to be fully loaded again.

Also undos some flaky tags done in 2018:
12e537593179cab96c27eb184ed7719184ea205f
c1c182d958a47c9d65da5f962890347f3f7cf3ba

As this might have been the real culprit.
If not, those should be re-added.

See #8348, #5390