Don't register schemes with QDesktopServices
On macOS there are two ways for us to be passed URIs to open, QFileOpen
events sent to the QGuiApplication, and the QDesktopServices
integration.

I think the commit https://codereview.qt-project.org/c/qt/qtbase/+/558406 made
it so if you have a QDesktopServices association for a scheme, you won't
see requests to open that scheme via QFileOpen events.

From the bug report, the QFileEvents happen when:

* open links from other applications. E.g. clicking on a link in the terminal or a mail program
* open -a /Applications/qutebrowser.app ...

The QDesktopServices integration in this case is so that we get a chance
to open all URLs in qutebrowser before passing them to the OS. For
example clicking a link outside of the web content (webinspector etc)
would otherwise open it in the OS' default browser.

The root issue looks to have been fixed in 6.8.3: https://codereview.qt-project.org/c/qt/qtbase/+/629664