Remove unneeded conditional
pylint says `Possible using variable 'parent_idx' before assignment. Which is true. Git blame shows that this conditional was added in a commit that just says "lint fixes".
Looking at the logic, the only tab that doesn't have a parent should be the tree root, and we shouldn't be acting on that because of the `if node.value is None` guard at the top of the loop. So hopefully this conditional isn't needed at all.
I'm not actually sure the logic is complete here anyway, this codeblock says it's inserting a hidden tab back into the tab widget. But it's just inserting it at parent_idx+1, shouldn't it be putting it in the correct place in the siblings list too? At least there are lots of other places in the codebase that do that.