Make :tab-give --recursive test more thorough
It was broken so that only one tab was being removed from the donor
window and this wasn't being caught in the test that was only looking at
the target window.

Using the "session should look like" test to make sure all the right
tabs are open, then the "tabs should be open" one to check the tab
hierarchy is right.