Move tree tabs from the tabMoved signal handler
Previously we moved tree tabs in the :tab-move command directly, which
let us operate on the tree as the first thing after a user invoke the
command and avoid any tree-naive code. But some tabs moves can come from
other sources, like dragging a tab which is currently handed by QTabBar.

This commit moves tree specific tab moving code to TreeTabWidget where
it's triggered from the tabMoved signal of QTabBar. Which lets us use
the same code for moves we initiate vs those initiated elsewhere.

This code is definitely less confident than the previous version. Since
we are acting asynchronously we have to use a bit of guesswork to figure
out what we are supposed to be doing, and then since we are moving tabs
recursively we have to be careful to avoid more move events we trigger
while wrangling children.

First we check if anything needs doing. If the code that triggered the
move event has already updated the tree structure, then we've got
nothing to do. We do that by comparing the tab at `to_idx` (which has
been moved already) with the tree node at `to_idx` (which may not have
been updated from the move yet). Having to keep the tab bar and tree
structure in sync so that they match up when the tree is traversed is a
bit awkward, but a number of other places already rely on it (including
`update_tree_tab_positions()` which is called from a number of places
and updates the QTabBar indexes based on the tree structure).

Then, after making sure we are not already in the middle of a move
operation with a recursion guard, we carefully apply the move operation
to the node that corresponds to the tab that already got moved. By
detaching the node from the tree and inserting it back into in the tree
alongside the node corresponding to the tab that is neighbouring the
moved tab.

After that we need to update the tab titles, so just fire off the catch
all `tree_tab_update()` method.