Add back check for moving tab to its own descendant
This operation is blocked because it would confuse the current single
case logic of tree tab moving. Currently we move a tab an all of it's
children at once (recursive tab moving). So detaching the whole tab
group and moving it to be under a tab in the group itself won't work.

We could support this operation, we would just have to add a special
case for it and decide what that should look like. Maybe:

1. move all the tabs between the current tab and the target tab to be
   under the target tab (or should it be making the current tab a
   sibling of the target tab?) and move the target tab up to be in the
   current tab's place
2. just move the current tab, not it's children

The tab moving code has moved into TreeTabWidget behind the tabMoved
signal. But we need to check for this case here if we want to raise a
nice CommandError. Otherwise we would have to complain and do nothing
(even though the tabBar would have already moved the current tab) or
complain and try to undo the move, which could be fraught (or could be
easier tbh, I haven't tried it).