Handle dragging a tree tab with the mouse
This commit provides logic for adjusting the tab tree after a user drags
a tab in the tab bar. It's currently quite brute force code that handles
each different case I could think of separately. Moving between
siblings, moving up or down a branch, and moving between tab groups.
It also handles tabs moving up the tab bar differently than tabs moving
upwards. Some cases are pretty simple in that we just swap the nodes and
their children, the cases where a node moves between trees are a little
more ticket.

The logic is largely informed by the animation the QTabBar does. The
user would already have seen the animation, so we are left figuring out
how to make sense of that in the tab tree. QTabBar just swaps two rows,
so that's our starting point. Then we adjust levels of tabs and
parent/child relationships.

There should be some opportunities to consolidate and reduce the line
count.