add Node.insert_child() method to simplify child list juggling
It was a common pattern in this tab dragging code to do:

    node_parent = node.parent
    node_siblings = list(node_parent.children)
    node_rel_idx = node_siblings.index(node)
    node_siblings.insert(node_rel_idx, other_node)
    node_parent.children = node_siblings

This inserts `other_node` as a sibling of `node`. But there is a lot of
line noise so it's hard to spot.

This commit extracts that out to a `insert_child()` method so that it
can be done in a single line.

Also changes a bunch of code doing manual manipulation of child lists to
use it. There are likely other places that can be updated too.