Switch to two lifecycle state timers
The previous implementation of the lifecycle timer had one timer that we
would change based on what state were were heading into. Because we
connected a new function to it every time we had to disconnect previous
connections before using it again which was a bit awkward, it raised a
TypeError if there were no connections, which isn't very specific.

This commit is an attempt to switch to using two timers, each with their
own callback.

I took a stateless/idempotent approach where the logic doesn't know what
state it might have been in previously so needs to stop any timers that
might be set each time, and needs to check if the timer it wants to set
is already active. Potentially the line count could be reduced by
encoding a bit more state in the logic. But hopefully it's simpler to
maintain like this.

Since I had need to (potentially) stop the timers at several points in
the logic, I pulled that out to a helper function which is where most of
the complexity is now. I tried to make the helper function a little more
dynamic by using a dict lookup to get the new timer and a loop to
reactive the old one. This didn't really add much, I think it's a
slightly higher line count, and probably a little harder to parse. The
previous implementation just looked a bit repetitive and hard to scan
over:

    to_start = delay = None
    if state == QWebEnginePage.LifecycleState.Frozen:
        to_start = self._lifecycle_timer_freeze
        self._lifecycle_timer_discard.stop()
        delay = config.val.qt.chromium.lifecycle_state_freeze_delay
    elif state == QWebEnginePage.LifecycleState.Discarded:
        self._lifecycle_timer_freeze.stop()
        to_start = self._lifecycle_timer_discard
        delay = config.val.qt.chromium.lifecycle_state_discard_delay
    elif state == None:
        self._lifecycle_timer_freeze.stop()
        self._lifecycle_timer_discard.stop()
    else:
        raise utils.Unreachable(recommended_state)

    if to_start and not to_start.isActive() and delay != -1:
        log.webview.debug(f"Scheduling recommended lifecycle change {delay=} {state=} tab={self}")
        to_start.start(delay)

I re-worked the logic of the `_on_recommended_state_changed()` changed
function a little. Not moving into the `Active` state is done immediately
instead of via a 0ms timer. So it's more like an guard clause and early
exit. I think it flows a little better now with all the decisions layed
out clearly and acted on one at a time instead of having the disabled
check right at the start but still proceeding through the function in
one case.