mypy: cast variable only on Qt5
We only have to cast these variables on Qt5. If we cast them on Qt6 mypy
warns of a redundant cast. If we ignore that warning mypy complains of a
redundant ignore on Qt5.

Add a conditional on `machinery.IS_QT5` and only cast it required.

But also move that conditional out to a utility function to avoid
duplicating code in the function doing the actual logic.

I'm re-using the `_T` generic TypeVar defined above. Not sure if it's
proper to re-use them or not. Doesn't python have a better way of
defining generics now?

I added the `valid-type` ignore because mypy was complaining:

    Variable "to_type" is not valid as a type

I'm not sure if there is some way to do it better, but `reveal_type()`
says that the call site is getting the right type back in the end.