Maybe move to pip as build frontend in makefile
It seems to me that installers are more likely to have pip installed
than build, so if we can get by using pip as the frontend in the
makefile, then we should.

pip does not have a way to build an sdist (tarball), so we'll have to
see if any of the packaging workflows depend on that. It does seem to
dump the source tree in dist/lib, but I'm not sure if that is part of
the API or not.