Fixed minor issue in configuration docs
The docs show an example for adding domain filtering for configuration options. However the example only matches the root of a domain rather than all pages on a domain which is for example, the default case when using the `tsh` shortcut to disable/enable javascript on a page.

(cherry picked from commit 1cbb6fccf0f52fd8baecf6f222fa53b5ff9fe2d0)